Fix Isu 31-4 comments
