+++ src/app/FarmerDashboard.tsx
@@ -9,22 +9,7 @@ import {
   Bell,
   Calendar,
 } from "lucide-react";
-
-interface Produce {
-  id: number;
-  name: string;
-  description: string;
-  farmer: string;
-  farmerLocation: string;
-  price: number;
-  unit: string;
-  quantity: number;
-  image: string;
-  category: string;
-  inSeason: boolean;
-  organic: boolean;
-  harvestDate: string;
-}
+import { Produce } from "./types";
 
 interface FarmerDashboardProps {
   produce: Produce[];
