+++ src/app/components/ProductCard.tsx
@@ -2,23 +2,7 @@
 import React from 'react'
 import Image from 'next/image'
 import { MapPin, Heart } from 'lucide-react'
-import AddToCart from './AddToCart'
-
-interface Produce {
-  id: number;
-  name: string;
-  description: string;
-  farmer: string;
-  farmerLocation: string;
-  price: number;
-  unit: string;
-  quantity: number;
-  image: string;
-  category: string;
-  inSeason: boolean;
-  organic: boolean;
-  harvestDate: string;
-}
+import { Produce } from '../types'
 
 interface ProductCardProps {
   product: Produce;
@@ -84,7 +68,12 @@ const ProductCard: React.FC<ProductCardProps> = ({ product, onAddToCart, onViewD
         </div>
 
         <div className="flex gap-2">
-          <AddToCart product={product} onAddToCart={onAddToCart} />
+          <button
+            onClick={() => onAddToCart(product)}
+            className="flex-1 bg-green-600 text-white py-2 px-4 rounded-lg hover:bg-green-700 transition duration-300 font-medium"
+          >
+            Add to Cart
+          </button>
           {onViewDetails && (
             <button
               onClick={() => onViewDetails(product)}
