+++ src/app/types/index.ts
@@ -1,0 +1,15 @@
+export interface Produce {
+  id: number;
+  name: string;
+  description: string;
+  farmer: string;
+  farmerLocation: string;
+  price: number;
+  unit: string;
+  quantity: number;
+  image: string;
+  category: string;
+  inSeason: boolean;
+  organic: boolean;
+  harvestDate: string;
+}
